cloud-functions/background/requirements.txt
# Function dependencies, for example:
google-cloud-storage
